Fix: GaugedBalancingPolicy: Unregister metrics on shutdownWork
If we don't, then then MetricsRegistry will continue holding onto old
gauges and meters, and when reporters like GraphiteReporter ask the
MetricsRegistry for all known metrics, it will get a bunch of orphaned
metrics that we never unregistered when the work unit was shutdown.
Querying the orphaned meters will just return and report 0 values, which
is mostly harmless, but querying the orphaned gauges will ask for things
that might no longer exist, which is potentially harmful. Further, the
orphaned gauges could hold references to otherwise orphaned objects and
cause memory leaks.