Continue work to reduce spurious claim cycles.
Along with the previous commit, this diff collapses the claim cycles
that would be triggered by incoming "nodeRemoved" events into a single
claim cycle. In local testing on a cluster of four Ordasity nodes and
2600 work units, this reduced failed claims following the death of a
node from over 600,000 to 2,565. It would be an understatement to say
that these changes will leave Ordasity apps and ZooKeepers much happier.