Resolve an issue introduced by work-claiming optimizations in version 0.4.9.
Ordasity 0.4.9 introduced an issue preventing work from being claimed if
new work unit znodes were added to the cluster. Specifically, the change
directed Ordasity nodes to only claim new work units if they were pegged to
the node. This behavior occured due to my error in confusing the "upsert"
and "insert" behavior of a ZooKeeper map. Apologies for this regression.