Set a watch for immediate handoff of work.
Currently, ordasity will schedule both the node handing off the work
unit and the new node accepting the work unit with a delay, after which
the new node will attempt to claim the work unit by creating an
ephemeral node in ZK. There is a slight race condition here, so often
times the handoff is delayed by a fixed delay time (of 2s) before the
node can be claimed.

This behavior has been modified to use a ZK watch on the ephemeral node
owned by the node handing off the work unit. When the watch fires, the
new node will more reliably be able to claim the work unit without
triggering the delay.